Don't try and delete session keyrirng. Fixes bug #355414
	* gnome-keyring-daemon-file.c: Don't try and delete session 
	keyrirng. Fixes bug #355414

svn path=/trunk/; revision=538

+++ gnome-keyring-daemon-file.c
@@ -370,9 +370,11 @@ remove_keyring_file_from_disk (GnomeKeyring *keyring)
 {
 	int res;
 
-	g_assert (keyring->file != NULL);
+	/* Cannot remove session or memory based keyring */
+	if (keyring->file == NULL)
+		return FALSE;
+
 	res = unlink (keyring->file);
-		
 	return (res == 0);
 }
